Take BasicEnum out of DaysOfWeek
